FossilOrigin-Name: 12d0a8859de0a9d823997cfeccc77bd572cb6d13

+++ src/where.c
@@ -3000,7 +3000,7 @@ static int codeAllEqualityTerms(
 
   /* Evaluate the equality constraints
   */
-  assert( zAff==0 || strlen(zAff)>=nEq );
+  assert( zAff==0 || (int)strlen(zAff)>=nEq );
   for(j=0; j<nEq; j++){
     int r1;
     pTerm = pLoop->aLTerm[j];
